- (snmplib/snmpv3.[ch]): Don't assume size_t is signed (Pekka Kytolaakso)   - N.B: This changes the interface of snmpv3_generate_engineID() slightly.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@3776 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmpv3.h
@@ -26,7 +26,7 @@ int     snmpv3_store(int majorID, int minorID, void *serverarg,
 u_long  snmpv3_local_snmpEngineBoots(void);
 int     snmpv3_clone_engineID(u_char **, size_t* , u_char*, size_t);
 int     snmpv3_get_engineID(u_char *buf, size_t buflen);
-u_char *snmpv3_generate_engineID(size_t *);
+u_char *snmpv3_generate_engineID(int *);
 u_long  snmpv3_local_snmpEngineTime(void);
 int     get_default_secLevel(void);
 oid    *get_default_authtype(size_t *);
